Make getDraggedBlocksOrigin private
